Address bugs and fix reported security issue (#3)
* Changed type of 'offset' parameter to use int instead
Removed unused NagerDateHttpClient parameter from Context.java
Remove unused parameter 'key'
Fixed ignored --version flag
Fixed GetLongWeekend ignores --available-bridge-days and --subdivision
Updated README.md to reflect Java 21 requirement and removed --key option


* Upgraded jackson version to 2.18.6 to patch GitHUb reported CWE-770


---------

+++ nager-date-api/src/main/java/com/github/flombois/services/v3/PublicHolidayV3Service.java
@@ -91,7 +91,7 @@ public interface PublicHolidayV3Service {
      * @return true if today is a public holiday, false otherwise
      * @throws NagerDateServiceException if an error occurs while communicating with the API
      */
-    boolean isTodayAPublicHoliday(CountryCode countryCode, String countyCode, byte offset) throws NagerDateServiceException;
+    boolean isTodayAPublicHoliday(CountryCode countryCode, String countyCode, int offset) throws NagerDateServiceException;
 
     /**
      * Checks whether today is a public holiday in the specified country, wrapped in an Optional.
@@ -106,7 +106,7 @@ public interface PublicHolidayV3Service {
      * @return an Optional containing {@code true} if today is a public holiday, {@code false} otherwise,
      *         or empty on failure
      */
-    default Optional<Boolean> getOptionalTodayAPublicHoliday(CountryCode countryCode, String countyCode, byte offset){
+    default Optional<Boolean> getOptionalTodayAPublicHoliday(CountryCode countryCode, String countyCode, int offset){
         try {
             return Optional.of(isTodayAPublicHoliday(countryCode, countyCode, offset));
         } catch (NagerDateServiceException e) {
